completed logic for matching pooled instance fleets
